+++ src/client/debugger/debugAdapter/main.ts
@@ -17,14 +17,14 @@ import { Disposable } from 'vscode';
 import { DebugSession, ErrorDestination, Event, logger, OutputEvent, Response, TerminatedEvent } from 'vscode-debugadapter';
 import { LogLevel } from 'vscode-debugadapter/lib/logger';
 import { DebugProtocol } from 'vscode-debugprotocol';
-import '../../../client/common/extensions';
+import '../../common/extensions';
 import { isNotInstalledError } from '../../common/helpers';
 import { IFileSystem } from '../../common/platform/types';
 import { ICurrentProcess } from '../../common/types';
 import { createDeferred, Deferred, sleep } from '../../common/utils/async';
 import { noop } from '../../common/utils/misc';
 import { IServiceContainer } from '../../ioc/types';
-import { AttachRequestArguments, LaunchRequestArguments } from './../types';
+import { AttachRequestArguments, LaunchRequestArguments } from '../types';
 import { CreateAttachDebugClient, CreateLaunchDebugClient } from './DebugClients/DebugFactory';
 import { BaseDebugServer } from './DebugServers/BaseDebugServer';
 import { initializeIoc } from './serviceRegistry';
