+++ src/test/debugger/misc.test.ts
@@ -17,7 +17,7 @@ import { DebugClientEx } from './debugClient';
 
 const debugFilesPath = path.join(__dirname, '..', '..', '..', 'src', 'test', 'pythonFiles', 'debugging');
 
-const EXPERIMENTAL_DEBUG_ADAPTER = path.join(__dirname, '..', '..', 'client', 'debugger', 'mainV2.js');
+const EXPERIMENTAL_DEBUG_ADAPTER = path.join(__dirname, '..', '..', 'client', 'debugger', 'debugAdapter', 'main.js');
 
 let testCounter = 0;
 const testAdapterFilePath = EXPERIMENTAL_DEBUG_ADAPTER;
